ユーザーデータの処理を外部に移動 #36
