+++ components/atomos/miniMapDiv/index.jsx
@@ -3,6 +3,7 @@ import React from "react";
 import L from "leaflet";
 import "../../../utils/extwms";
 import EventListener from "react-event-listener";
+import UserData from "../../../utils/userData";
 
 class MiniMapDiv extends React.Component {
   myMap = null;
@@ -29,17 +30,7 @@ class MiniMapDiv extends React.Component {
     );
 
     // ユーザーデータ取得(cookieから持ってくる)
-    const userData = { user_id: "", access_token: "" };
-    const r = document.cookie.split(";");
-    r.forEach(function(value) {
-      const content = value.split("=");
-      content[0] = content[0].replace(" ", "");
-      if (content[0] == "user_id") {
-        userData.user_id = content[1];
-      } else if (content[0] == "access_token") {
-        userData.access_token = content[1];
-      }
-    });
+    const userData = UserData.getUserData();
 
     const mainLayer = L.TileLayer.wmsHeader(
       "https://pascali.info-mapping.com/webservices/publicservice/WebmapServiceToken.asmx/WMSService?TENANTID=21000S",
