+++ components/organisms/infoTypeSelector/index.jsx
@@ -2,6 +2,7 @@ import "./infoTypeSelector.scss";
 import Router from "next/router";
 import React from "react";
 import InfoTypeItem from "../../molecules/InfoTypeItem";
+import UserData from "../../../utils/userData";
 
 const BoarDiv = () => (
   <div className="boar-div select-div">
@@ -48,49 +49,8 @@ class InfoTypeSelector extends React.Component {
     this.state = {
       selected: null
     };
-    // ユーザーデータ取得(cookieから持ってくる)
-    const userData = { user_id: "", access_token: "" };
-    if (process.browser) {
-      const r = document.cookie.split(";");
-      r.forEach(function(value) {
-        const content = value.split("=");
-        content[0] = content[0].replace(" ", "");
-        if (content[0] == "user_id") {
-          userData.user_id = content[1];
-        } else if (content[0] == "access_token") {
-          userData.access_token = content[1];
-        }
-      });
-    } else {
-      return;
-    }
 
-    // 本番：ユーザーIDの１文字目からユーザーを識別
-    // const userDepartment = userData.user_id.substr(0, 1).toUpperCase();
-    // テスト環境：ユーザーIDから識別
-    // どうして仕様に則ったユーザーIDじゃないの…
-    let userDepartment;
-    switch (userData.user_id) {
-      case "tyousa":
-        userDepartment = "T";
-        break;
-      case "yuugai":
-        userDepartment = "U";
-        break;
-      case "shityouson":
-        userDepartment = "S";
-        break;
-      case "trap":
-        userDepartment = "W";
-        break;
-      case "pref":
-        userDepartment = "K";
-        break;
-      default:
-        userDepartment = null;
-        break;
-    }
-
+    const userDepartment = UserData.getUserDepartment();
     // userDepartmentに応じて表示するものを変更する
     switch (userDepartment) {
       case "T":
