+++ components/templates/Detail/index.jsx
@@ -10,6 +10,7 @@ import Header from "../../organisms/header";
 import Footer from "../../organisms/footer";
 import RoundButton from "../../atomos/roundButton";
 import FooterAdjustment from "../../organisms/footerAdjustment";
+import UserData from "../../../utils/userData";
 
 class Detail extends React.Component {
   state = {
@@ -24,17 +25,7 @@ class Detail extends React.Component {
       Router.push("/map");
       return;
     }
-    const userData = { user_id: "", access_token: "" };
-    const r = document.cookie.split(";");
-    r.forEach(function(value) {
-      const content = value.split("=");
-      content[0] = content[0].replace(" ", "");
-      if (content[0] == "user_id") {
-        userData.user_id = content[1];
-      } else if (content[0] == "access_token") {
-        userData.access_token = content[1];
-      }
-    });
+    const userData = UserData.getUserData();
 
     const receiptNumber = Math.floor(Math.random() * 100000);
     const data = {
