+++ utils/session.js
@@ -1,6 +1,6 @@
 import Router from "next/router";
 import cookies from "next-cookies";
-import "./statics";
+import UserData from "./userData";
 
 export default class SessionManager {
   static isLogin(ctx) {
@@ -18,19 +18,7 @@ export default class SessionManager {
   }
 
   static logout(document) {
-    const userData = { user_id: "", access_token: "" };
-
-    const r = document.cookie.split(";");
-
-    r.forEach(function(value) {
-      const content = value.split("=");
-      content[0] = content[0].replace(" ", "");
-      if (content[0] == "user_id") {
-        userData.user_id = content[1];
-      } else if (content[0] == "access_token") {
-        userData.access_token = content[1];
-      }
-    });
+    const userData = UserData.getUserData();
 
     const receiptNumber = Math.floor(Math.random() * 100000);
     const data = {
